Implement bypassing the cache experiment
If a large data structure is produced and won't be used right
away, it might evict data from the cache that will be needed
soon. To resolve this situation, we can use non-temporal write
operation, which makes the new content directly written to memory to
bypass the cache.

In this experiment, I reproduce the bypassing cache test to
compare normal write operations with non-temporal write operations
to access large data structure, both in sequential and random
access patterns. The results are obtained by measuring the time
taken to initialize a 3000 x 3000 matrix.