Implement matrix multiplication experiment
Accessing the L1 cache significantly influences program performance.
We can improve locality and align the code and data to optimize it.

In this experiment, I implement four methods provided in the paper
about matrix mulplication, with one original version and three
improved versions for optimizing the use of the loaded cache lines.
I use `perf` to measure the number of CPU cycles they spend.